+++ service/users.js
@@ -19,16 +19,23 @@ class UserService {
     return userRepository.addUser(userWithHashedPassword);
   }
 
-  removeUser(userName) {
-    userRepository.deleteUser(userName);
+  async removeUser(userName) {
+    const user = await userRepository.getUserByUserName(userName);
+
+    if (!user) {
+      throw new InexistentItem('The user with this username does not exists.');
+    }
+
+    return userRepository.deleteUser(userName);
   }
 
   async changeUser(user) {
     const updated = await userRepository.updateUser(user);
 
     if (!updated[0]) {
-      throw new InexistentItem('Username does not exists.');
+      throw new InexistentItem('The user with this username does not exists.');
     }
+
     return updated[1][0];
   }
 }
